refactor: 로그인 상태 변경로직 중복 제거
-  HTTP Error의 handle함수 추가
- App을 생성해서 export하는 방식으로 변경
